Program compiles using the Makefile compiler options.

+++ hw2/src/reformat.c
@@ -15,6 +15,7 @@
 #include <stdlib.h>
 #include <ctype.h>
 #include <string.h>
+#include<stdio.h>
 
 #undef NULL
 #define NULL ((void *) 0)
@@ -193,7 +194,7 @@ char **reformat(const char * const *inlines, int width,
     if (end - *line < affix) {
       sprintf(errmsg,
               "Line %d shorter than <prefix> + <suffix> = %d + %d = %d\n",
-              line - inlines + 1, prefix, suffix, affix);
+              (int) (line - inlines + 1), prefix, suffix, affix);
       goto rfcleanup;
     }
     end -= suffix;
